# contracts/openapi/core-api-v1.json
{
  "openapi": "3.1.0",
  "info": { "title": "NEXUS WORLD Core API", "version": "1.3.0" },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 80 },
          "password": { "type": "string", "minLength": 1, "format": "password" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "contractVersion",
          "accessToken",
          "tokenType",
          "expiresIn",
          "refreshToken",
          "refreshExpiresIn"
        ],
        "properties": {
          "contractVersion": { "const": "v1" },
          "accessToken": { "type": "string", "minLength": 1 },
          "tokenType": { "const": "Bearer" },
          "expiresIn": { "type": "integer", "minimum": 1 },
          "refreshToken": { "type": "string", "minLength": 1 },
          "refreshExpiresIn": { "type": "integer", "minimum": 1 }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "minLength": 1 }
        }
      },
      "CurrentUser": {
        "type": "object",
        "additionalProperties": false,
        "required": ["contractVersion", "userId", "username", "roles"],
        "properties": {
          "contractVersion": { "const": "v1" },
          "userId": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "roles": {
            "type": "array",
            "uniqueItems": true,
            "items": { "enum": ["VIEWER", "ANALYST", "OPERATOR", "ADMIN"] }
          }
        }
      },
      "Problem": {
        "type": "object",
        "required": ["title", "status", "detail"],
        "properties": {
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/api/v1/platform/status": {
      "get": {
        "operationId": "getPlatformStatus",
        "responses": {
          "200": {
            "description": "Core and AI service contract status",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/platform-contract-v1.json#/$defs/PlatformStatus" }
              }
            }
          },
          "503": { "description": "AI service contract unavailable" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Short-lived access token",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "400": { "description": "Invalid request" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshAccessToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Rotated access and refresh token pair",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "401": { "description": "Invalid, expired, revoked, or reused refresh token" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" } }
          }
        },
        "responses": {
          "204": { "description": "Token family revoked and current access token blacklisted" },
          "401": { "description": "Invalid access or refresh token" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user claims",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CurrentUser" } }
            }
          },
          "401": { "description": "Missing, invalid, or expired access token" }
        }
      }
    },
    "/api/v1/sources": {
      "post": { "operationId": "createSource", "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/CreateSource" } } } },
        "responses": { "201": { "description": "Source registered", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Source" } } } }, "400": { "description": "Invalid source" }, "409": { "description": "Duplicate source key" } } }
    },
    "/api/v1/sources/{id}": {
      "get": { "operationId": "getSource", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Source", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Source" } } } }, "404": { "description": "Source not found" } } }
    },
    "/api/v1/evidence": {
      "post": { "operationId": "createEvidence", "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/CreateEvidence" } } } },
        "responses": { "201": { "description": "Evidence registered", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Evidence" } } } }, "400": { "description": "Invalid evidence" }, "404": { "description": "Source not found" } } }
    },
    "/api/v1/evidence/{id}": {
      "get": { "operationId": "getEvidence", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Evidence", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Evidence" } } } }, "404": { "description": "Evidence not found" } } }
    },
    "/api/v1/assumptions": {
      "post": { "operationId": "createAssumption", "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/CreateAssumption" } } } },
        "responses": { "201": { "description": "Assumption registered", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Assumption" } } } }, "400": { "description": "Invalid assumption" }, "409": { "description": "Duplicate assumption key" } } }
    },
    "/api/v1/assumptions/{id}": {
      "get": { "operationId": "getAssumption", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Assumption", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/Assumption" } } } }, "404": { "description": "Assumption not found" } } }
    },
    "/api/v1/provenance-links": {
      "post": { "operationId": "createProvenanceLink", "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/CreateProvenanceLink" } } } },
        "responses": { "201": { "description": "Provenance link registered", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/ProvenanceLink" } } } }, "400": { "description": "Exactly one origin is required" }, "404": { "description": "Origin not found" }, "409": { "description": "Duplicate link" } } },
      "get": { "operationId": "getProvenanceLinks", "security": [{ "bearerAuth": [] }], "parameters": [
          { "name": "subjectType", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "subjectId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Ordered provenance links", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/ProvenanceLink" } } } } } } }
    },
    "/api/v1/provenance-links/{id}": {
      "get": { "operationId": "getProvenanceLink", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Provenance link", "content": { "application/json": { "schema": { "$ref": "../schemas/provenance-contract-v1.json#/$defs/ProvenanceLink" } } } }, "404": { "description": "Provenance link not found" } } }
    }
  }
}